Move integration tests into integration package
