+++ complete-project-documentation.yaml
@@ -78,8 +78,8 @@ directory_structure:
               - PropertyUtilTest
               key_methods:
               - checkGuessPropertyName
-            IntegrationTest.kt:
-              purpose: Test file IntegrationTest.kt verifying behavior.
+            integration/IntegrationTest.kt:
+              purpose: Test file integration/IntegrationTest.kt verifying behavior.
               key_classes:
               - IntegrationTest
               key_methods:
@@ -227,8 +227,8 @@ directory_structure:
               - testWrapper
               - after
               - performance
-            integrationStubs.kt:
-              purpose: Test file integrationStubs.kt verifying behavior.
+            integration/integrationStubs.kt:
+              purpose: Test file integration/integrationStubs.kt verifying behavior.
               key_classes:
               - SettingsStub
               - ISettingsState
